pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]

[project]
name = "crs"
version = "0.1.0"
description = "Compliance Recommendation System: NIST 800-53 control mapping with OSCAL evidence and auditor feedback."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "scikit-learn>=1.3",
  "pyyaml>=6.0",
  "matplotlib>=3.7",
  # Optional: enable if you add embedding baseline
  # "sentence-transformers>=2.6",
  # "torch>=2.2",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
testpaths = ["tests"]
